Cleaned up the folder. Reimplemented interfaces as per the UML
