+++ .flake8
@@ -1,0 +1,3 @@
+[flake8]
+ignore = E402
+max-line-length = 120
