+++ grasp_path_planner/scripts/Models/models_old/DQN_Model_SimpleSim_30k.zip
