+++ grasp_path_planner/scripts/Models/models_old/DQN_Model_local_CARLA_10k.zip
