+++ grasp_path_planner/scripts/Models/models_old/DQN_Model_local_minClamp_simple_10k.zip
